requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
aiosqlite==0.18.0
asgiref==3.8.1
asyncpg==0.28.0
attrs==23.2.0
bcrypt==4.1.2
Brotli==1.1.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
commonmark==0.9.1
Django==5.0.3
djangorestframework==3.15.1
feedparser==6.0.11
frozenlist==1.4.1
geographiclib==2.0
geopy==2.4.1
idna==3.6
Jinja2==3.1.3
MarkupSafe==2.1.5
multidict==6.0.5
mysqlclient==2.2.4
packaging==24.0
prompt-toolkit==3.0.43
pycparser==2.21
python-olm==3.2.16
questionary==1.10.0
requests==2.31.0
ruamel.yaml==0.17.40
ruamel.yaml.clib==0.2.8
sgmllib3k==1.0.0
SQLAlchemy==1.3.24
sqlparse==0.4.4
urllib3==2.2.1
uWSGI==2.0.24
wcwidth==0.2.13
wheel==0.43.0
yarl==1.9.4
redis==5.0.3

